requirements.txt
pandas
sentence-transformers
chromadb
openai
requests
ollama
pyarabic
tqdm
python-dotenv